Make logging less verbose at the Info level.
Notifications from the API server (Add, Delete, Update) are now
logged at the Debug level.

Details of the monitor loop are also at the Debug level when all is
well, for example when there are no panics, child is running, etc.

Otherwise, messages previously at the Debug level are now at the
Trace level -- logging to trace and debug what the controller is
doing.