Wait for queues and workers to shut down before exiting.
This has exposed k8s issue #59822 when we use client-go v6.0.0 for
compatibility with k8s 1.9, when shared informers are stopped as
part of queue shutdown:

https://github.com/kubernetes/kubernetes/issues/59822

This upgrades to use of client-go v6.0.1, but evidently only has the
effect that the problem is intermittent. Occasionally there are panics
due to double channel close. The guarantee of sequential delivery
by a shared informer may also not be honored.

We go forward with these issues, since 1.9 is already out of support,
and will probably start supporting newer k8s versions in the near
term.

Ref #6